requirements.txt
colorama
Pillow
